requirements.txt
colorama
Pillow
